refactor: 강의관리 페이지 비즈니스/렌더링 로직 분리
